change gaji data type to int

+++ src/assignment/Karyawan.java
@@ -4,13 +4,13 @@ public class Karyawan {
 
     private int nip;
     private String nama;
-    private float gaji;
+    private int gaji;
     private float lamaKerja;
 
     public Karyawan() {
     }
 
-    public Karyawan(int nip, String nama, float gaji, float lamaKerja) {
+    public Karyawan(int nip, String nama, int gaji, float lamaKerja) {
         this.nip = nip;
         this.nama = nama;
         this.gaji = gaji;
@@ -33,11 +33,11 @@ public class Karyawan {
         this.nama = nama;
     }
 
-    public float getGaji() {
+    public int getGaji() {
         return gaji;
     }
 
-    public void setGaji(float gaji) {
+    public void setGaji(int gaji) {
         this.gaji = gaji;
     }
